+++ CHANGELOG.md
@@ -30,6 +30,10 @@ We have encountered transient lambda service errors in our integration testing. 
 - **CUMULUS-1319**
   - Fixed a bug where granule ingest times were not being stored to the database
 
+- **CUMULUS-796**
+  - Added production information (collection ShortName and Version, granuleId) to EMS distribution report
+  - Added functionality to send daily distribution reports to EMS
+
 ## [v1.13.0] - 2019-5-20
 
 ### PLEASE NOTE